fixed migration error
