+++ sugahfree/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
-from .models import User, Input
+from .models import User
+from myapi.models import Input
 
 admin.site.register(User)
 admin.site.register(Input)
